Added & fixed print methods

+++ complex.h
@@ -6,7 +6,7 @@
 bool is_space(char sym);
 
 int CharToInt(char sym);
-int StrToInt(char *str, size_t start);
+int StrToInt(char *str, int start);
 
 //typedef double complex Item;
 typedef double Element;
@@ -23,7 +23,7 @@ Complex complex_summ(Complex num1, Complex num2);
 Complex complex_composition(Complex num1, Complex num2);
 void complex_print(Complex num);
 
-Complex complex_set_value(char *str, size_t start);
+Complex complex_set_value(char *str, int start);
 bool complex_is_zero(Complex num);
 
 #endif
